requirements.txt
# Database
psycopg2-binary==2.9.9

# HTTP client
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

argparse==1.4.0

numpy==1.26.0
pydantic==2.4.2 